feature: slice search utility functions
As the "SearchString" function from the "sort" package returns the index
in which the target string should be inserted, a little helper code is
needed to actually return a boolean indicating the search result.

Also, the helper "Sort*" function is there to also provide a easy way of
searching a target string in a slice which might be unsorted.