refactor: move source validation to "service" package
I was getting "circular imports" issues when trying to import the "dao"
package in the "model" package —because the model package then imports
the "dao" package back, thus creating the aforementioned issue". Moving
the validation to a different package will allow us to use both the
"dao"s and the "model"s to properly validate the incoming data.