feature: add initial migration from the current state of the db
This migration is a snapshot of the current schema of the database. In
order not to make it clash with it when it gets applied, some safeguards
are implemented which check the existence of the tables and the
corresponding foreign keys.

Therefore, the tables only get created when they don't exist, and the
same thing goes with the foreign keys.

Althought not everything gets checked: the sequences and constraints
associated with a table are assumed to be non existent if the table
itself doesn't exist.