refactor: run Miniredis in a non-default port
Apparently Miniredis is causing some issues with people who are running
a full Redis instance locally. It seems that turning off the Redis
instance makes the tests pass again, so changing the port on which
Miniredis runs should do the trick.

Instead of using the "RunT" function of Miniredis and requesting a
random port, a hardcoded one is used instead. This is to avoid having
the one in a million chance of getting the default Redis port as the
random port. Hopefully this will avoid head scratching to the developer
who has to experience tests failing once, and then passing without any
issue in the next run.