refactor: start and stop the Miniredis instance at the package level
Instead of having to start and stop Miniredis every time, which consumes
resources and bloats the code, the instance can be run at the package
level and leave to the tester the responsability of creating a new
instance if they so desire.

It is true that the state is shared among the tests, but for now this is
not a problem.