refactor: return the external tenant instead of the tenant id
The Rails app returned the "external_tenant" of the related tenant of a
given source, so this Go version should basically do the same. In order
to avoid having to fetch the entire table, a "Join" is performed, but
GORM seems not to be able to map stuff directly to an inner struct
—which is a fair limitation, to be honest—.

That's why I'm using a half way solution for this: instead of looping
through the rows and manually mapping the fields, I let GORM handle that
at the expense of fetching all the columns of the joined table, and
letting it make a "LEFT JOIN" instead of an "INNER JOIN".