refactor: use a "for ... range" loop instead of a manual one
Due to using the "make" syntax wrong, I was hitting some "index out of
bounds" panics when doing a for range. Basically I was swapping the
length and the capacity:

make([]string, 0, len(r.sources))

Instead of the following one:

make([]string, len(r.sources))

That's explains the error, since in the for loop I was trying to access
the "sourcesIds[i]" element. It also explains why I retorted to using
a more "manual" loop.