refactor: don't stringify the marketplace token when using Vault
If the token gets serialized as a string and then assigned, the double
quotes of the resulting JSON get escaped, thus returning a very weird
JSON to the clients.

Instead, if the "BearerToken" struct is directly given, once the above
levels serialize the "extra" field properly by creating a proper JSON
object out of whatever there might be.