refactor: generate JSON content dynamically
Using JSON files made it very difficult to modify the hardcoded dates,
which was why the status listener tests had so many functions: the dates
had to be updated from the database every time the tests ran.

Since the database contents are just what we have in the fixtures, it
occurred to me that it might be easier just to fetch the database
contents, and generate the JSON model for what the status listener
should be genearting.

This reduces a step in the tests, since we go from:

1. Parse all the JSON.
2. Fetch the latest database timestamps.
3. Modify the timestamps in the JSON content.
4. Compare.

To

1. Fetch the latests database contents.
2. Build the expected JSON.
3. Compare.

Also, reestructuring the code like this makes it easier to refactor the
tests in the future.