refactor: check manually if the relation exists before inserting
The reason to remove "FirstOrCreate" in this code block is that the
behaviour for how the "RowsAffected" gets updated has changed.

In the previous version of the Gorm and/or driver libraries, a
"FirstOrCreate" operation which fetched an object instead of creating it
returned a "RowsAffected" with a value of 0, whilst in the newer version
of the libraries, it returns 1.

Therefore, we cannot really rely on "RowsAffected" to guess if the
object was either created or it was found, to then return a proper
response with it.

This is why the code has been changed to a "manual" check instead of
relying on the "FirstOrCreate" function.