fix: not able to connect to reset or setup the database
When the database didn't exist, the service was still trying to connect
to a non-existing database, which would make the program exit and not
even reach the "reset" or "setup" code.

In order to avoid that problem, if either of those flags is passed, a
connection is opened to the maintenance "postgres" database, to be able
to perform those operations, since it is not possbile to drop a database
from whithin the connection that is connected to it.