refactor: use a custom "authentications" struct for creating the table
Gorm was creating FKs for the "authentications" database table, but in
none of the environments we do have such FKs. There doesn't seem to be
a way of telling Gorm to don't do that without modifying the original
model, and since that's not viable —I don't want to modify the model's
definitions because of the tests—, using an anonymous struct seems to be
the way of avoiding those FKs.