refactor: adapt tests to the database backed authentications
Prior to these changes, the authentications were always expected to be
returned and serialized with the Vault structure, including its string
IDs.

Now, in order to make the unit tests work, we need to simulate that
Vault is online, so that the expected payloads are returned. This also
coincides to when the integration tests are run with Vault online.

For the database tests, as the ID changes, different payloads need to be
checked when running the integration tests.