refactor: disable status listener tests for database backed auths
The tests of the status listener code fail because essentially we are
loading the JSON in memory and changing the timestamps so that we can
check if the status listener outputs the payload we expect it to output.

The problem with that approach is that in the case of database backed
authentications, things like the "Vault path" of an authentication don't
make sense, and therefore the expected structure should be different.

I initially attempted at creating a set of tests for the status listener
and the database backed authentications, but I found it very difficult
to do so. That's why I opened a PR[1] suggesting a refactor that would make
things simpler for this particular use case.

Since things are moving forward, we agreed to skip the status listener
tests and add them for the database backed authentications once the
refactor is in place.

So what I've done is simply make the status listener tests work as if we
only supported Vault, because that way they work all the time. Once the
refactor of the status listener tests is done we will be able to
properly modify the tests to include the database backed authentications
to the test mix.

[1]: https://github.com/RedHatInsights/sources-api-go/pull/139