fix: "too many connections" error in tests
Due to the large number of database tests we've got, and how we openned
a new connection for each one of them, the database was starting to
complain about having too many active connections at once.

The test utils have been refactored to use a single connection for the
tests.