fix: bulk delete fetches all the resources
The "Find" function of Gorm can take a slice of IDs so that the function
adds a "WHERE ... IN" condition automatically. The catch is that when
the slice is empty Gorm doesn't add any where condition, which ends up
resulting in a "SELECT * FROM" query, which fetches the whole table.