fix: app auths are tried to be fetched by the wrong column
The "application authentications" DAO was trying to fetch the objects by
looking at the "authentication_uid" column, when a list of
authentications were given to the function that was failing.

Since we are using the database datastore for now, the
"authentication_uid" column does not exist yet, and therefore the query
fails every time.

This error popped up on the status listener since that's the one which
is using the function in question.