refactor: improve logging and error messages
Until now we usually logged the error that the functions outputted, and
we ended up returning that same error to the user. The problem is that
generally those errors are "unrecoverable" from the user's standpoint:
they can't do much about it.

Therefore I think it makes sense to return a generic error message along
with an "internal server error" response to let them know that it wasn't
their fault, and that they should report the error whenever it is possible.

On the other hand, in our case we should be logging the error with as
much context as possible, and that's why I changed the log messages for
the marketplace code.