fix: marketplace token doesn't get serialized to "extra.marketplace"
Apparently I forgot to include the token in the "extra.marketplace" key
in one case, and the test I wrote for it didn't catch that because it
was expecting the token in the "extra" key.

I've changed the implementation so that the token always gets serialized
to the "extra.marketplace" key, and so that the test looks at having the
token in the "extra.marketplace" key.