refactor: improve log messages when the migrations fail
This commit changes the signature of the function, since we were only
returning the error to log it afterwards. I thought that, in that case,
it would be better to spare that return and to log the error directly.

The commit also improves the error messages, so that it is clearer why
migrations are failing. For example, when Redis was not present, the
error message would look like as if the database were not running, when
in reality the "connection error" came from the Redis client.