refactor: improve consistency when forwarding headers
In that function we were looking for specific headers on the request to
forward them, except in one case: the "psk-account", which is a context
variable set once the headers have been parsed.

For the sake of being consistent, I've changed it to check for the exact
header the "psk-account" is coming from.