fix: "failed to pull identity from request" when using PSK and EBS
When using the PSK and the EBS account number headers, the tenancy
middleware overwrote the "identity" context value with an "Identity"
struct, instead of an "XHRID" one. That caused the "bulk create"
operation to not be able to pull the expected "XHRID" struct from the
context, and therefore return the aforementioned error.