feature: migration for updating the duplicated tenants' subresources
This migration grabs all the duplicated tenants —the ones that have the
same "external_tenant" or "org_id" column values— and picks the first
one as the main one.

Then, for the other duplicates of this tenant, it updates all the
subresources' —application authentications, applications, authentications,
endpoints, sources and source rhc connections— tenant id to the main tenant.

Finally, it updates the duplicated tenants with an easily identifiable
value for the "external_tenant" and "org_id" columns.