fix: "schema_migrations" legacy table not being created on initial schema
Apparently I copied and pasted the "IF NOT table exists" statement with
the "ar_internal_metadata" value instead of the legacy
"schema_migrations" value. This fixes that, although it doesn't affect
to the already populated databases or the new ones when running the
"delete the old active record table" migration.