refactor: replace Kafka manager with stateless functions
The Kafka manager felt like it was designed as a singleton, but in our
codebase all the readers and writers were being closed after using them.
This means that we did not have a long running, single Kafka reader or
single Kafka writer to attend to all the requests/responses.

That is why, in my effort to decouple the "kafka/" module from the
"config/" module as much as possible, I thought that maybe we could get
rid of the Kafka manager in favour of simpler functions which just
return the reader and the writer.

However, some of the manager's functions have been kept since they were
useful for certain stuff, like deferring the closure of the resources.