refactor: validate the ID using regex
By just using the integer parsing function, we were allowing numbers
which had a sign on them, which are a valid integer number, but which
would make very weird looking valid URLs.

The regex validation should take care of that. Also, the parse integer
function is left untouched because that would help us in the "number is
too big" edge case.