refactor: centralize tenancy lookups on the identity struct
Until now, we were "individually" looking up the tenants in three ways:

1. By creating an identity struct from the EBS account number.
2. By creating an identity struct from the OrgId.
3. By passing the identity struct directly.

Since the third way includes the two above, I thought it might be a good
idea to just use the identity struct for everything, and to forget about
pulling specific data for the tenancy.

This will make the "x-rh-identity" header the favoured in case we
receive it along the "ebs account number" and "orgid" headers.