feature: not null constraint and default value for "availability_status"
This migration adds a not null contraint and a default value for all the
tables which have an "availability_status" column. The default value is
"in_progress", a value that was agreed as the default one by the team.

It also takes care of all the current NULL or empty values and sets them
to the default value too.