feature: application user for the sources application
The idea is to have two users in the sources application: a fully
privileged one, which will solely be used to perform database migrations
and the seeding.

The application user, on the other hand, is granted just the required
CRUD permissions on the tables for the application to work.

There are two goals:

* Restrict what the database user can do when operating normally.
* Pave the way to enable row level security via session variables in a
  future update.