feature: run SonarQube scanner containerized
This change containerizes the keystore creation and runs the scanner on
a container too.

The pipeline's runner has a Java 1.8 as the Java runtime environment,
and even though the scanner has its own JRE embedded, it might
eventually complain if we feed it a keystore generated with an old JRE.

That's why the keystore creation is containerized. And since we're at
it, we also containerized running the scanner.