feature: translate EBS account numbers to "orgIds"
After checking for untranslated EBS accout numbers, I found that there
are quite a few that are not translated. However, I was able to manually
translate them with the translation tool, so I am suspecting that we
might get rid of quite a bit of untranslated tenants if we perform a
second run of this migration.

After this, I'd like to check for untranslated EBS account numbers again
to discuss a proper move for these.