feature: add translation endpoints
We were initially going to perform another migration to translate the
leftover tenants which don't have an associated "OrgId" value, but after
a stand up meeting we decided upon creating a couple of endpoints to be
able to do it on demand.

These endpoints will help at:

1. Fetching all the "translatable" tenants.
2. Initiating the "EBS account number to OrgId" process.

Of course, these endpoints go on the internal routes, so they will only
be able to be hit from inside the cluster.