refactor: tenancy middleware sets the tenant data from the database
Instead of just forwarding the incoming headers, this change ensures
that the "EBS account number" and the "OrgId" are set in the context
using the contents from the database.

This comes from one issue we have been having with the notifications
service: the service was erroring because we were just sending the
incoming "EBS account number" header, even though we had the associated
"OrgId" on the database.

So we have decided to simply use the incoming headers as a validation
—make sure that an EBS account number and/or OrgId are present—, and
then use the database contents to send the headers to the rest of the
services or to the Kafka queues.