Revert "feature: log the events to help debugging issues"
After checking the logs on Kibana, I saw that they don't show much
information. Basically the inner structs are shown as references, and
the other data isn't as helpful to look at.

Also, the vast majority of the times "Raise" is called, if an error
occurs, it gets logged by the caller function, so these log statements
end up just duplicating the messages and making everything harder to
follow.

This reverts commit d1f95029007ab6195d7fc8bc286c032a00943408.