Availability check 'scheduled' URL flag
A new 'scheduled' flag is accepted for the availability flag which is passed to applications over HTTP in the JSON payload. A followup PR for the background scheduler will be created to set this flag for all scheduled checks. Applications can now decide to perform immediate actions or increase checking priority when the flag is missing (or false), meaning the request was not made from the scheduled job.