feature: support "external ID" field for Cloud Meter roles
Cloud Meter is adopting the "external ID" field in their roles to make
use of the suggested feature from AWS to restrict the access to third
parties to AWS accounts.

Therefore, we need to update the Superkey metadata for the application
so that the new roles are created with that new field.

RHCLOUD-35160