refactor: move encryption initialization to "main"
Superkey depends on Sources API Go, and when updating its dependencies,
launching the application with the Clowder configuration enabled was
causing panics. It turned out that it was due to how encryption is
initialized, which pulls the Sources API Go config and attempts to read
it.

Obviously, when this happened in the Superkey application, it caused
panics because the configuration for the Superkey worker and the Sources
API Go application is different.