fix: the "background worker" and the "status listener" reported as down
Prometheus is not able to reach the "background worker" and the "status
listener" because those two services don't have a "metrics" endpoint.
This is causing some alerts to be fired because the aforementioned pods
are marked as "down", not "up". By creating a simple metrics' exporter
for all of the services we will ensure that they show up as "up and
running" for Prometheus.

RHCLOUD-38530