refactor: RBAC base path creation and RBAC client creation
We get the RBAC service's URL from an environment variable but we should
really get it from Clowder's configuration file, in case the service's
hostname or port changes.

Also, since the legacy RBAC client requires the whole base path
including the "/v1" suffix, and we might need to send requests to the
"/v2" API soon, I refactored the way the RBAC client gets created so that
we can easily extend or refactor it in the future.

RHCLOUD-39709