chore: adjust golangci linter
The PR enables all the linters that do not cause any current error
messages with the codebase, but that we want to keep them enabled for
any new patches.

It also annotates the linters we would like to enable to improve our
code's quality, but since they require code changes they are kept
disabled to be able to progressively enable them while making those
changes.