fix: the user does not have permissions to create a file
I initially modified the Dockerfile to remove the part where "777"
permissions were given a directory from which the Sources application
ran.

Since there were no comments about why that was happening, the
application could not launch in the stage environment because "it lacked
permissions to create the RDS' CA file", which comes from Clowder's
configuration.

Instead of restoring the previous behavior, I have opted for creating a
user without privileges and giving that user a directory which the user
owns, so that the binary can freely create the RDS' CA file coming from
Clowder without having to give the "777" permissions to a directory.