feature: log bulk creation requests' results and status messages
When resources were created via the "Bulk request" endpoint, we did not
log the resources that got created, which made debugging issues a bit
harder.

The same thing happened with the status listener messages: we did not
get any information of the incoming messsages unless it was the raw
Kafka message, which many times we are not interested in.

RHCLOUD-40425