fix: Sources is returning "405 — Method not allowed" errors
The upgrade of the "Platform Go Middlewares" library changed the
"System" struct to now be a pointer inside the "Identity" struct, and
the "emptiness" check was not being correctly performed. This caused
some requests to be incorrectly identified as being authenticated with
certificates, and since not all of them met the requirements, those
requests were getting responded with "405" responses.

RHCLOUD-42357